+++ src/recover.h
@@ -8,7 +8,7 @@
 #ifndef FTI_RECOVER_H_
 #define FTI_RECOVER_H_
 
-int FTI_CheckFile(char *fn, long fs, char* checksum);
+int FTI_CheckFile(char *fn, int32_t fs, char* checksum);
 int FTI_CheckErasures(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
         FTIT_topology* FTI_Topo, FTIT_checkpoint* FTI_Ckpt,
         int *erased);
